Added Region enum
This enum encapsulates the mapping between numbers (`u8`)
and the regions.

A `Region` can be parsed from a string, can be displayed
(or converted to `String` using `to_string()`) and it's
easy to get the corresponding region id from a `Region` instance
using `region as u8`.

The advantage is that this can be used as an argument type
which means the compiler can check that no invalid region id
can be passed to a function/method that expects a valid Region.