Improvements to `Target` enum, etc...
The `Target` enum is used to determine what kind of
input the user provided as first argument of the cli.

This change better uses Rust's enum by encapsulating
the postcode String or the `Region` values in the
`Target` discriminants.

This means when parsing the input we also get the user-provided
value within the enum.

Because of this `match` statements are nicer and can
extract the postcode/region, e.g.

```Rust
match target {
    Target::Region(region) => {
        // here region was destructred out of the Target
        // use `region` ...
    }
    _ => { ... }
}
```