don't panic in `to_tuple()`
`to_tuple()` was using unwrap instead of handling
the error, e.g. return an `Err(ApiError)` in this
case.

AFAICT `to_tuple()` is only used on date strings
coming from the API so it should be fine, but it
still makes sense to handle potentiall parsing errors.

NOTE: The `?` operator implicitly converts the given
value into the return error type via `From`/`Into`.
That's what the `#[from]` attribute is doing:
it's implementing the `From` trait to convert a
`chrono::ParseError` into an `ApiError`.