Make requests concurrently
When requesting data for different date ranges make these requests
concurrently instead of making them sequentially.

Broadly speaking this is what's happening:
- spawn a Tokio task for each of these requests.
  - each task will return a `Result<Vec<IntensityForDate>>`
- wait for them to return using `try_join_all()`
  - this returns a `Result<Vec<Result<Vec<IntensityForDate>>>>`
- if it's an error (e.g. some of the task panicked) we return it
  using `?`
- we then have a `Vec<Result<Vec<IntensityForDate>>>`. One element
  for each of the requests' results.
- we "invert" that and turn it into `Result<Vec<Vec<IntensityForDate>>>`
  - either we return the first error or unwrap the list of tuples
- we now have a `Vec<Vec<IntensityForDate>>`, one element per response
  and each element is a `Vec<IntensityForDate>` (the list of tuples
  with date/intensity)
- we finally flatten this into a simpler `Vec<IntensityForDate>`

This is a bit more convoluted than I'd like but I hope it makes sense.

Some of the types in the variable declarations are there mainly
for readability sake given the levels of nesting but some of them
are there to help Rust figure out how to `collect()`.

**NOTE**: I'm still not 100% happy with this and I do wonder if the
unwrapping/error handling could be done more elegantly - possible.
Also, potentially there may be a way of doing this in a single bigger
step, I may experiment with it further to see if it turns out to be
a bit more readable.