Make it explicit crate needs Rust 1.80+
Although the crate is in theory build for Rust Edition 2021/`1.56` in
practice when I tried to run the test using 1.56 it fails.

- `chrono` needs [1.57+](https://crates.io/crates/chrono/0.4.31)
- `clap` needs [1.70+](https://crates.io/crates/clap/4.4.8)
- `clap_builder` needs [1.74+](https://crates.io/crates/clap_builder/4.5.19)

We're now using the amazing `std::sync::LazyLock` but this was stabilised
in Rust 1.80: https://doc.rust-lang.org/stable/std/sync/struct.LazyLock.html

Adding `rust-version` to the `Cargo.toml` just makes this explicit rather
than implicit.